podcast-shorts/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "podcast-shorts"
version = "0.1.0"
description = "AI-powered podcast shorts generation platform"
requires-python = ">=3.11"
dependencies = [
    # LLM orchestration
    "langchain>=0.3.0",
    "langgraph>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langgraph-checkpoint-sqlite>=2.0.0",

    # Trend research tools
    "tavily-python>=0.5.0",
    "pytrends>=4.9.0",
    "google-api-python-client>=2.100.0",
    "tweepy>=4.14.0",

    # Media generation
    "elevenlabs>=1.0.0",
    "lumaai>=0.2.0",

    # Video processing
    "moviepy>=2.0.0",
    "pysrt>=1.1.0",

    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",

    # Core utilities
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "httpx>=0.27.0",
    "structlog>=24.0.0",
    "sse-starlette>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/podcast_shorts"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
